more cleanup, remove dead code
